Interim Release Code (#11)
* Moved Name to Composite-level

Reason: every component has a name.  Easier to address at the Composite-level and not have to retype everything.

* CompositeProject finished.

Incorporated DocumentSupportType in every Composite class to match.

* CompositeGroup finished.

All according to the diagram.

* Everything but CompositeClass

Documentation updated, arguments renamed, and functionality added.

* Brought CompositeClass up to speed.

Some of the stuff outlined in the static class diagram is different.  The implementation in this commit was done with intention to correct the diagram at a later time.

* Fix: Lint.

Always what catches us in the end.