Frontend Editing + Other Fixes (#72)
* Documented main.component for Reference

It's going to make it easier to make the modifications needed for the frontend.

* Added support for descriptions.

Fixes #60.
Users can now modify descriptions.

* Added ability to select other documentation.

Fixes #69, for the most part.  Nothing dynamic since it works the same way as we do for languages for the time-being.

* Lint Commit

* Fixed Documentation Format Issue

Semantic error on my end.

* Preliminary Form Builder

* Re-modified Buttons

Now they check and look nice.

* Added Ability to Overwrite Descriptions

* Linting + Documentation

* HTML + CSS + Linting

Minor adjustments to make the mainline edit page look nice.  Still handles well.

* Set Right What Once Went Wrong

Accidentally resolved the merge conflict improperly.  Fixed it.